Code ui for rock paper scissors
So admittedly I haven't been adhering to the "commit early, commit
often" rule, so here is what I have so far. There are still many bugs,
but the buttons invoke the playRound function, which add to the correct
score and announce the winner (or if it's a tie). I'm currently working
on getting a "try again" button to pop up, and the game play to cease
after either the human or computer reaches five points. For now the
game will play infinitely. Also there is still a lot of extra code
commented out, what can I say, sometimes I hoard. Bad habits, but I
intend on correcting them. Baby steps. Also I intend on making more
frequent commits on this project and others. That's all for now, byeee!